pine64-pinebookPro: Force 0x8000 for SPI TPL
I don't really actually care *where* it should be loaded, but I do know
that putting the TPL at 0x8000 and the board looking at 0x6000 is not
going to work well.

Since there's not a real convention for that location, let's just force
it to what I know worked previously.

Hopefully this also gives us room to spare if, for any reason, the SPL
grows much more.