rockchip-rk3399: Implement SPI installer image
The build has been reworked to assemble the output from multiple
derivations, rather than hijack the main Tow-Boot build.

Since we're now using the image-builder infra, we can more easily
compose additional things, like here, we are composing an SPI installer
by appending a new partition to the base image.

Some of that infra will be moved in the future, since it is likely to be
the same for many boards where we're flashing with `sf`.