image-builder: rewrite makeGPT with sfdisk/sgdisk
Ugh... This is really odd in the end, but this is meant to allow us to
*then* use the feature where sfdisk knows how to reduce the number of
entries in the GPT.

This is because the "holey" GPT strategy fails on some OS like Fedora,
which ends up manipulating the GPT even if it does not need to, or asked
not to!!! (Might not be limited to Fedora.)

So the next commit will introduce "number of entries" parameter to
reduce the GPT size, which should allow creating a discrete partition
rather than hiding the firmware.

In the end, a discrete partition is probably for the best.