tow-boot: Add support for better TUI menu
This builds up onto PDCursesMod, a curses implementation.

This, in turn, provides a more ergonomic framework to build a rich TUI,
which can be used to build much more than this simplistic menu.

This is not a 1:1 replacement for the built-in ANSI escape sequences
based menu. It is also bringing a more ergonomic menu building
foundations, which allows building the menu dynamically from the default
distro_bootcmd data.

This is the kind of changes Tow-Boot aims to be a incubator for, and
foster.