pine64-pinebookPro: Add display reset patch
This is coming from the author of the patch set bringing display
support. The hack is not acceptable in U-Boot upstream.

This is the kind of patches we import to *pragmatically* improve board
support, as long as it is always built on top of the most recent
mainline, and is not an invasive change.