tow-boot: Use pause command on interactive boot failure
This gives the user more control during the boot process.

This only happens on *interactive* boot, e.g. when using the menu. As
otherwise this would SUCK if the first tried boot option failed!

This also gives them time to read the boot failure, in addition to
allowing them to skip to the menu more quickly. Beforehand they had to
sit tight for 5 seconds.