spi-installer: Guard installation with board identifier
This helps prevent the dangerous situation where you could write the
wrong firmware to the wrong board.

While writing the firmware for another family of board is likely not an
issue, writing a firmware that looks valid enough to the boot rom for
the board is likely to put the user in a pickle.

Fixes #3