spi-installer: Harden against bad flashes
This is done by splitting the write in a head and tail of the firmware.

We first remove the SPI Flash from the boot chain by making it invalid
(zeroing out the start).

Then we write the end of the firmware.

Finally we write the "head".

Hopefully any unexpected failure (e.g. sudden power loss) happens
before the head is written, this way the user is not "bricked".