support/docker: Add build shim
This adds a shim allowing end-users to invoke a containerized Nix
instance, in a manner compatible to using `nix-build`.

```
 $ support/docker/build.sh -A pine64-pineA64LTS
```

This is equivalent to

```
 $ nix-build -A pine64-pineA64LTS
```

The semantics vary a little. *Only* the `result` symlink is resolved and
replaced with the store path contents. This means that using
`--out-link` is not supported at this time, and multiple build outputs
won't work as expected.

This should be sufficient though.

Co-authored-by: Linus Heckemann <git@sphalerite.org>