pine64-pinephonePro Work around eMMC issues with UMS
It seems the current patch series works fine for the tested scenarios,
including the ones I had previously found out.

But it's not enough for UMS to work like it did previously.

The following messages will be observed:

```
*** ERROR: Can't read GPT header ***
find_valid_gpt: *** ERROR: Invalid GPT ***
mmc fail to send stop cmd
```

Though using the configuration just added seems to work around the
issues, unclear why. These were the options the patch author was testing
against.

This could come from insufficient backporting. An alternative would be
to go back to the previous workaround, and disable the feature that is
the root cause of issues.