pine64-pinephonePro: fix ALS interrupt pin
On this device, GPIO4_D3 is used as the interrupt pin for the
proximity/ambient light sensor. However, even though the datasheet
indicates it's configured as an input by default, the boot ROM sets this
pin as an output, causing issues when probing the sensor in linux.

This adds a patch for making sure this line is configured properly on
the PPP, and reverts some previous DT changes (those break
`dm_lookup_by_name()` on RK3399, as the driver expects the gpio node
names to include the bank number right before '@').