pine64-pinephonePro: Use appropriate OPP
The vendor requested that the maximum frequencies and maximum voltage
respect the RK3399-T datasheet.

This applies to all Pinephone Pro, even those without an RK3399-T, be it
the unseen RK3399s, or the standard RK3399. This is because while the
SoC may handle it fine, the design was validated with the lower
frequencies, and lower thermals. The LCD panel may operate out of range
when the device is clocked higher than designed for.

See also:

 - https://gitlab.com/pine64-org/linux/-/merge_requests/36

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>