tow-boot/smbios: Add fallback sysinfo DT nodes
These nodes are pre-filled with the Tow-Boot board-specific info. This
is correct enough as a fallback. We should strive to upstream discrete
entries to all boards where sysinfo should be filled.

This will require "generic" builds to include different DT entries for
the different boards in the correct `*-u-boot.dtsi` files.