installer: Install both bootrom partitions
In issue !76 support to installing to eMMC boot partitions is requested,
however currently, only a single partition is being installed. This can
be problematic if the BootROM for whatever reason decides to boot from
the second boot partition. The SPL could also in the future make
decisions based on this (corruption detection, and since SPL should
match the actual U-Boot, triggers a reboot after a switch).

To support this use-case, install the software into both boot
partitions.

Flexibility might be required in the future, if certain distro's/boards
decide they want to keep the second boot partition for other purposes,
but that should be discouraged and should be unlikely, as the second
boot partition is for redundancy or A/B support.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>