installer: Ensure Boot configuration is setup
The dual boot partition support in the eMMC does not require any BootROM
support. This because the eMMC spec mandates the so called 'boot
partition selection'.

This means, that you tell the eMMC chip, which partition to 'pretend' to
boot from. IOW, the BootROM tries to read from the 'default user
partition' but the eMMC chip actually serves the data from the other
partitions.

That does not mean the BootROM cannot do better things, such as using
'high-speed boot', where the eMMC chip actually streams the data in the
pre-configured way.

Both options are possilbe, and this patch enables this as an optional
feature. Without setting the needed variables, nothing gets done and
setup.

In short, this means that:

mmcBootAccess: The type of boot access desired, 'stream' mode, or 'normal
mmcBootAck: the ACK command is sent to the host after CMD0 was sent
mmcBootBusWidth: The bus-width used during 'stream' access mode
mmcBootBusWidthReset: The bus-width after a reset of the eMMC chip
mmcBootMode: The boot bus mode
mmcBootPart: Boot partition if not 1 or 2 (e.g. 0 or 7).

This means that to start a BootROM that does not know or care about eMMC
boort partitions, gets told via `partconf <dev> 0 2 0` that any read it
does will be comming from eMMC boot partition 2.
Alternativily `partconf <dev> 1 1 1` sets the eMMC chip to 'stream'
data, with an ACK, from eMMC boot partition 1. This requires special
support from the loader, but can be significantly faster.

For the specific details see the U-Boot documentation [0].

[0] https://u-boot.readthedocs.io/en/latest/usage/cmd/mmc.html

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>