boards/RPi: switch to the u-boot aarch64 combined image
The current version of the raspberryPi-aarch64 image only works on
a couple of RaspberryPi boards: 3b and 4b.

Adding support for other boards can be achieved by using u-boot's
`rpi_arm64_defconfig`, which supports:

 - Raspberry Pi 3b
 - Raspberry Pi 3b+
 - Raspberry Pi 4b
 - Raspberry Pi 400
 - Raspberry Pi CM 3
 - Raspberry Pi CM 3+
 - Raspberry Pi CM 4
 - Raspberry Pi CM 4s
 - Raspberry Pi zero 2 w

This however requires shipping the .dtb files along with the overlays
inside the shared image, which this commit also adds.

Unfortunately, the raspberrypifw package is now outdated, so rather
than depending on nix OS packages, we also vendor our own copy of it.