boards/RPi: make sure the USB ports are enabled on the CM4
Tow-boot's user interface requires USB support to interact with the
boot configuration menu, which is unfortunately disabled by default
on the CM4, as can be seen in page 7 of the CM4 datasheet[1]:

    The firmware disables the USB interface by default to save power.
    In recent versions of Raspberry Pi OS (Bullseye) it is
    automatically enabled by the otg_mode=1 setting in the config.txt
    file. If you are using a different OS, or an older version of
    Raspberry Pi OS, you will need to add this to config.txt to enable
    the USB interface.

In my experience on a CM4, setting otg_mode=1 was not sufficient to get
my keyboard to work in both Linux or Tow-Boot, while using
worked in all cases `dtoverlay=dwc2,dr_mode=host`. This is why I opted
to use this setting.

As per this comment[2], it also seems to apply to the CM4S.

[1] https://datasheets.raspberrypi.com/cm4/cm4-datasheet.pdf
[2] https://www.jeffgeerling.com/comment/31588#comment-31588