doc: Fixup site links for the devices namespace
The main issue here is mismatched expectations. GitHub expects refs to a
dir to show `README.md`, like you would `dir/` to get `index.html`.
Instead we are producing discrete html pages per boards. So the links
don't match.

The fix here is to transform those links to refer to the page.