tow-boot/builder: Drop EXTRAVERSION, always
It is assumed we follow the release, and the only SUBLEVEL release they
did was to fix the EXTRAVERSION being left from the last RC.

To remove complexity with making `2023.07.02` expose itself as
`2023.07`, we'll instead ignore the problem for now.

if/when they do a SUBLEVEL release that warrants exposing we are using
it, we'll re-evaluate how we handle that.