modules/tow-boot: Implement tagged releases
This is necessary since we want to keep track of changes between actual
releases, so e.g. have `-rcX` in `development` during the lifecycle of
the upgrade cycle, but being able to send fixes to the U-Boot codebase.

By using tags for every time we change the U-Boot codebase, we can more
easily `diff` between our changes.